requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-docx
tiktoken
python-dotenv